+++ libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: ddecourt <ddecourt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/24 21:30:11 by ddecourt          #+#    #+#             */
-/*   Updated: 2021/04/03 15:20:43 by ddecourt         ###   ########.fr       */
+/*   Updated: 2021/04/03 23:16:16 by ddecourt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ typedef	struct	s_struct
 int		ft_display_string(const char *s, t_flags flags);
 int		ft_display_char(char c, t_flags flags);
 int		ft_display_int(int n, t_flags flags);
-int		ft_display_unsigned(int n, t_flags flags);
+int		ft_display_unsigned(unsigned int n, t_flags flags);
 int		ft_display_octal(int n);
 int		ft_display_hexa(unsigned int n, t_flags flags);
 int		ft_display_big_hexa(unsigned int n, t_flags flags);
@@ -45,7 +45,7 @@ void		ft_checkflag(char current, t_flags *flags, va_list args);
 t_flags		ft_init_flags(t_flags flags);
 int				ft_write_width(char c, int max, int size);
 int				ft_handle_flags_hexa(int size, t_flags *flags);
-char		*ft_convert(unsigned int n, char *base);
+char		*ft_convert(unsigned long int n, char *base);
 char		*ft_strchr(const char *s, int c);
 void		ft_putchar(char c);
 void		ft_putnchr(const char *s, int n);
